server/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
httpx
python-dotenv
groq
requests